Fix outcome config if every subset has been removed but one.
Until now, if the end condition was hit in the main loop then
the 'current' configuration was returned as a result. However, if
every subset was removed but one then that one subset, the
'next_config' would be the correct outcome instead the 'current_config'
which contains 2 subsets yet. The patch fixes this mistake.