Add wrapper for logging with additional levels and log methods
The commit introduces a wrapper for the standard logging package,
which clones all of its public interface and extends it with
* two new levels: `TRACE` and `DISABLE`,
* a dictionary `levels` that contains all log level names and their
  numeric values, and
* a `trace` method that gets set for all loggers returned by the
  wrapped `getLogger`.

Instead of using `import logging`, one can use
`from picire import logging` and benefit from the extended
functionality.

The CLI already makes use of it by lowering the logging level of
cache content dump from DEBUG to TRACE (because of its potentially
huge size).