Change config splitters to return list of slices
When splitting up a config (a list) into N pieces, splitters used
to generate N subsets, i.e., brand new lists containing some of the
elements of the original config. With this change the splitters
only return the boundaries of the subsets (slice objects), and an
actual subset is only computed when needed (by slicing the config
list with the appropriate slice object). This should improve both
performance and memory usage.