Ensure killing the still alive processes in _abort
The commit uses the psutil.Process.terminate when first trying to
stop the processes, since from psutil >5.4 its exitcode is non-zero
on Windows as well, which makes errors and graceful terminations
distinguishable. After this, we iterate through the processes that
weren't stopped after this terminate call and forcefully kill them.