No `test_pattern` argument for `tester_class`
Theoretically, tester classes may not necessarily use files to
perform tests, so the concept of test file pattern does not
necessarily apply to all of them. Moreover, tester classes should
be able to construct a path from the `config_id` (plus some extra
information like the root of a work directory and a file name to be
used for the tests) if there is a need for test files.

Therefore, `picire.call` does not pass `test_pattern` to tester
class initializers anymore, but ensures that `config_id`s are
unique. Moreover, `SubprocessTest` expects `work_dir` and
`filename` arguments (which can be determined before `picire.call`)
instead of `test_pattern` (which can only be determined inside
`picire.call`), and uses them to construct the test path when
called.